% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Excel2C.R
\name{Excel2C}
\alias{Excel2C}
\title{Excel-Style Hexavigesimal (\code{A-Z}) to C-Style (\verb{0-9; A-P})}
\usage{
Excel2C(x)
}
\arguments{
\item{x}{\link[base]{character} scalar or \link[base]{vector},
only letters \code{A-Z} and \code{a-z}.}
}
\value{
Function \code{\link[=Excel2C]{Excel2C()}} returns a
\link[base]{character} \link[base]{vector}.
}
\description{
Convert Excel-style hexavigesimal (\code{A-Z}) to C-style (\verb{0-9; A-P}).
}
\keyword{internal}
