% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Excel2int.R
\name{Excel2int}
\alias{Excel2int}
\title{Excel-Style Hexavigesimal (\code{A-Z}) to \link[base]{integer}}
\usage{
Excel2int(x)
}
\arguments{
\item{x}{\link[base]{character} scalar or \link[base]{vector},
only letters \code{A-Z} and \code{a-z}.}
}
\value{
Function \code{\link[=Excel2int]{Excel2int()}} returns an
\link[base]{integer} \link[base]{vector}.
}
\description{
Convert Excel-style hexavigesimal (\code{A-Z}) to \link[base]{integer}.
}
\keyword{internal}
