% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Excel2int.R
\name{hexavigesimalExcel}
\alias{hexavigesimalExcel}
\alias{Excel2int}
\alias{Excel2C}
\alias{int2Excel}
\title{Excel-Style Hexavigesimal (A to Z)}
\usage{
Excel2int(x)

Excel2C(x)

int2Excel(x)
}
\arguments{
\item{x}{\link[base]{integer} scalar or \link[base]{vector} for function \link{int2Excel}.
\link[base]{character} scalar or \link[base]{vector} for functions \link{Excel2C} and \link{Excel2int},
which consists of (except missingness)
only letters \code{A} to \code{Z} and \code{a} to \code{z}.}
}
\value{
Function \link{Excel2int} returns an
\link[base]{integer} \link[base]{vector}.

Function \link{Excel2C} returns a
\link[base]{character} \link[base]{vector}.

Function \link{int2Excel} returns a
\link[base]{integer} \link[base]{vector}.
}
\description{
Convert between decimal, C-style hexavigesimal (\code{0} to \code{9}, \code{A} to \code{P}), and Excel-style hexavigesimal (\code{A} to \code{Z}).
}
\keyword{internal}
