% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/int2Excel.R
\name{int2Excel}
\alias{int2Excel}
\title{\link[base]{integer} to Excel-Style Hexavigesimal (\code{A-Z})}
\usage{
int2Excel(x)
}
\arguments{
\item{x}{\link[base]{integer} scalar or \link[base]{vector}}
}
\value{
Function \code{\link[=int2Excel]{int2Excel()}} returns a
\link[base]{character} \link[base]{vector}.
}
\description{
Convert \link[base]{integer} to Excel-style hexavigesimal (\code{A-Z}).
}
\keyword{internal}
